+++ src/core/Vector2D.java
@@ -1,27 +1,27 @@
 package core;
 
 public class Vector2D {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Vector2D(int x, int y) {
+    public Vector2D(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 }
